testings/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
Deprecated==1.2.18
docstring_parser==0.16
exceptiongroup==1.3.0
fastapi==0.115.12
geographiclib==2.0
geopy==2.4.1
google-adk==1.1.1
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.95.1
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.18.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
numpy==2.0.2
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
wrapt==1.17.2
zipp==3.22.0
